♻️ refactor: 패키지구조 수정
