+++ src/main/java/com/example/Veco/domain/assignee/repository/AssigneeQueryDsl.java
@@ -1,7 +1,7 @@
 package com.example.Veco.domain.assignee.repository;
 
 import com.example.Veco.domain.assignee.entity.Assignee;
-import com.example.Veco.domain.mapping.MemberTeam;
+import com.example.Veco.domain.mapping.entity.MemberTeam;
 import com.example.Veco.global.enums.Category;
 
 import java.util.List;
