+++ src/main/java/com/example/Veco/domain/assignee/repository/AssigneeQueryDslImpl.java
@@ -2,7 +2,7 @@ package com.example.Veco.domain.assignee.repository;
 
 import com.example.Veco.domain.assignee.entity.Assignee;
 import com.example.Veco.domain.assignee.entity.QAssignee;
-import com.example.Veco.domain.mapping.MemberTeam;
+import com.example.Veco.domain.mapping.entity.MemberTeam;
 import com.example.Veco.global.enums.Category;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
