+++ src/main/java/com/example/Veco/domain/mapping/repository/MemberTeamRepository.java
@@ -10,7 +10,7 @@ public interface MemberTeamRepository extends JpaRepository<MemberTeam, Long> {
 
     List<MemberTeam> findAllByMemberIdInAndTeamId(List<Long> memberId, Long teamId);
 
-    Optional<MemberTeam> findByMemberId(Long memberId);
+    List<MemberTeam> findByMemberId(Long memberId);
 
     Optional<MemberTeam> findByMemberIdAndTeamId(Long memberId, Long teamId);
